This is a program that generates a magic square
This program generates a Square of a specified size inputted by the
user and tests if the generated square is magic. A magic square is a
square with the numbers 1-side^2 nonrepeating where the sum of each
row, column, and diagonal are equal. The side variable is the number of
elements in each row and column. In order for the sums to be equal
each sum must equal the Magic constant which is calculated by the sum
of all numbers (1-side^2)/side.